Split Program.cs into multiple files
